Flyway demonstration
This commit uses Flyway to generate the schema rather than letting
Hibernate create the schema based on conventions.

The only thing that is necessary is to create a `V1__Description.sql`
file in `src/main/resources/db/migration` and it will be invoked to
initialize the schema.

This commit also adds support for a persistent database. We've chosen
MySQL in this example. It is easy to switch from one database to
another using a profile.